refactor(gasi): 프로시저 추출
