+++ apps/gasi/src/routes/auth.ts
@@ -1,3 +1,4 @@
+import { RegisterUserRequestSchema } from "@request/specs";
 import { z } from "zod";
 import { kakaoAuthorize } from "../auth/providers.js";
 import { p } from "../trpc.js";
@@ -7,3 +8,5 @@ export const kakao = p
   .query(async ({ input, ctx }) => {
     return await kakaoAuthorize(input.redirectUrl ?? `${ctx.baseUrl}/callback`, input.code);
   });
+
+export const register = p.input(RegisterUserRequestSchema).mutation(() => {});
